Make version appear first in sarif_json_schema
The version property should appear first in resulting json object
despite json properties being unordered. This allows tools to quickly
assess if they support the version. It is not guaranteed by atd(gen)
that the fields are output in the same order, but it happens to be the
case with the current release. This could break in the future.