Property bags (#14)
* propertyBag is an object, do not derive ord

The propertyBag is an object with a further constraint that, if present,
the `tags` property is a list of strings. This extra constraint is not
enforced.

"ord", a compare function, is not derived because Yojson.Safe.compare
does not exist. The value of an order is also not clear to me.

A test is added that does a roundtrip of a sarif output taken from a
Semgrep test case.

* Add remaining Semgrep tests